requirements.txt
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
pytesseract==0.3.10
pygame==2.1.2
pywin32==305
     